move LookUpTxtTest to new package
